+++ fabric/src/main/java/me/drex/antixray/fabric/AntiXrayFabric.java
@@ -2,7 +2,9 @@ package me.drex.antixray.fabric;
 
 import me.drex.antixray.AntiXray;
 import me.drex.antixray.util.Platform;
+import me.lucko.fabric.api.permissions.v0.Permissions;
 import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.server.level.ServerPlayer;
 
 import java.nio.file.Path;
 
@@ -10,6 +12,11 @@ public class AntiXrayFabric extends AntiXray {
 
     public AntiXrayFabric() {
         super(Platform.FABRIC);
+    }
+
+    @Override
+    public boolean canBypassXray(ServerPlayer player) {
+        return Permissions.check(player, "antixray.bypass");
     }
 
     @Override