+++ forge/src/main/java/me/drex/antixray/forge/AntiXrayMod.java
@@ -1,11 +1,7 @@
 package me.drex.antixray.forge;
 
 import me.drex.antixray.AntiXray;
-import me.drex.antixray.util.Platform;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.loading.FMLPaths;
-
-import java.nio.file.Path;
 
 @Mod(AntiXray.MOD_ID)
 public class AntiXrayMod {